+++ src/services/openSearchService.ts
@@ -3,7 +3,10 @@ import { MyContext } from '../context';
 import { OpenSearchWork, WorkType } from '../types';
 import { awsConfig } from '../config/awsConfig';
 import { prepareObjectForLogs } from '../logger';
-import { createOpenSearchClient, OpenSearchConfig } from "../datasources/openSearch";
+import {
+  createOpenSearchClient,
+  createOpenSearchServerlessClient, OpenSearchConfig, OpenSearchServerlessConfig
+} from "../datasources/openSearch";
 import {
   OpenSearchRe3DataRecord,
   Re3DataRepositoryRecord,
@@ -87,9 +90,11 @@ interface OpenSearchRe3DataHit {
 
 export class OpenSearchService {
   private client: Client;
+  private serverlessClient: Client;
 
   constructor() {
     this.client = createOpenSearchClient(awsConfig.opensearch as OpenSearchConfig);
+    this.serverlessClient = createOpenSearchServerlessClient(awsConfig.opensearchServerless as OpenSearchServerlessConfig);
   }
 
   public async findWorkByIdentifier(reference: string, context: MyContext, doi: string | null | undefined, maxResults: number): Promise<OpenSearchWork[]> {
@@ -263,7 +268,7 @@ export class OpenSearchService {
 
     let response: unknown;
     try {
-      response = await this.client.search({
+      response = await this.serverlessClient.search({
         index: 're3data',
         body: {
           size: uris.length,
@@ -331,7 +336,7 @@ export class OpenSearchService {
           },
         };
 
-      response = await this.client.search({
+      response = await this.serverlessClient.search({
         index: 're3data',
         body,
       });
@@ -387,7 +392,7 @@ export class OpenSearchService {
         },
       };
 
-      response = await this.client.search({
+      response = await this.serverlessClient.search({
         index: 're3data',
         body,
       });